@@ -1,8 +1,8 @@
 export enum ServerRequestUrls {
     TASK = 'http://localhost:8080/tasks',
-    TASK_TYPE = 'http://localhost:8080/',
+    TASK_TYPE = 'http://localhost:8080/task-type',
     ATTEMPT = 'http://localhost:8080/attempts',
-    PUPIL_ATTEMPT = 'http://localhost:8080/pupil-attempts',    
+    PUPIL_ATTEMPT = 'http://localhost:8080/pupils-attempts',
     MISTAKE = 'http://localhost:8080/mistakes',
     RULE = 'http://localhost:8080/rules',
     SUBJECT = 'http://localhost:8080/subjects',
