@@ -1,7 +1,6 @@
-
-import { HttpClient }         from '@angular/common/http';
-import { Injectable }         from '@angular/core';
-import { Observable }         from 'rxjs';
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
 import { Group } from './group';
 
 @Injectable({
@@ -15,23 +14,38 @@ import { Group } from './group';
     }
 
     getGroup(id: number): Group {
-        var response = this.http.get<Group>(this.url + "/" + id);
+        const response = this.http.get<Group>(this.url + '/' + id);
         return this.parseToGroup(response);
     }
 
+    getFaculties(facultyUrl: string): Array<string> {
+        const response = this.http.get<Array<string>>(this.url + '/' + facultyUrl);
+        return this.parseToListFaculties(response);
+    }
+
+    getCourses(faculty: string, coursesUrl: string): Array<number> {
+        const response = this.http.get<Array<number>>(this.url + '/' + faculty + '/' + coursesUrl);
+        return this.parseToListCourses(response);
+    }
+
+    getGroupsByFacultyAndCourse(faculty: string, course: number): Array<Group> {
+        const response = this.http.get<Array<Group>>(this.url + '/' + faculty + '/' + course);
+        return this.parseToListGroups(response);
+    }
+
     getGroups(): Array<Group> {
-        var response = this.http.get<Array<Group>>(this.url);
+        const response = this.http.get<Array<Group>>(this.url);
         return this.parseToListGroups(response);
     }
 
     saveGroup(group: Group) {
-        var response = this.http.post<Group>(this.url, group);
+        const response = this.http.post<Group>(this.url, group);
         return this.parseToGroup(response);
     }
 
     deleteGroup(id: number) {
-        var response;
-        this.http.delete(this.url + "/" + id).subscribe(item => {
+        let response;
+        this.http.delete(this.url + '/' + id).subscribe(item => {
             response = item;
         }, error => {
             console.error(error);
@@ -40,10 +54,10 @@ import { Group } from './group';
     }
 
     private parseToGroup(observable: Observable<Group>): Group {
-        var group = new Group();
+        const group = new Group();
         observable.subscribe(item => {
             group.id = item.id;
-            group.cource = item.cource;
+            group.course = item.course;
             group.faculty = item.faculty;
             group.groupNo = item.groupNo;
             if (item.pupils) {
@@ -55,20 +69,44 @@ import { Group } from './group';
         return group;
     }
 
-    private parseToListGroups(observable: Observable<Array<Group>>): Array<Group> {
-        var pupils = new Array<Group>();
+    private parseToListFaculties(observable: Observable<Array<string>>): Array<string> {
+        const faculties = new Array<string>();
         observable.subscribe(response => {
             response.map(item => {
-                var group = new Group();
+                faculties.push(item);
+            });
+        }, error => {
+            console.error(error);
+        });
+        return faculties;
+    }
+
+    private parseToListCourses(observable: Observable<Array<number>>): Array<number> {
+        const courses = new Array<number>();
+        observable.subscribe(response => {
+            response.map(item => {
+                courses.push(item);
+            });
+        }, error => {
+            console.error(error);
+        });
+        return courses;
+    }
+
+    private parseToListGroups(observable: Observable<Array<Group>>): Array<Group> {
+        const pupils = new Array<Group>();
+        observable.subscribe(response => {
+            response.map(item => {
+                const group = new Group();
                 group.id = item.id;
-                group.cource = item.cource;
+                group.course = item.course;
                 group.faculty = item.faculty;
                 group.groupNo = item.groupNo;
                 if (item.pupils) {
                     group.pupils = item.pupils;
                 }
                 pupils.push(group);
-            })
+            });
         }, error => {
             console.error(error);
         });
