@@ -1,10 +1,10 @@
-import { PupilAttempt } from "../pupil-attempt/pupilAttempt";
+import { PupilAttempt } from '../pupil-attempt/pupilAttempt';
 import { Rule } from '../rule/rule';
 
 export class Mistake {
     id: number;
     pupilAttempt: PupilAttempt;
     rule: Rule;
-    rowNo: number;
-    colNo: number;
+    wordNo: number;
+    symbolNo: number;
 }
