@@ -1,7 +1,6 @@
-
-import { HttpClient }         from '@angular/common/http';
-import { Injectable }         from '@angular/core';
-import { Observable }         from 'rxjs';
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
 import { PupilAttempt } from './pupilAttempt';
 
 @Injectable({
@@ -15,36 +14,40 @@ import { PupilAttempt } from './pupilAttempt';
     }
 
     getPupilAttempt(pupilId: number, attemptId: number): PupilAttempt {
-        var response = this.http.get<PupilAttempt>(this.url + "/" + pupilId + "/" + attemptId);
+        const response = this.http.get<PupilAttempt>(this.url + '/' + pupilId + '/' + attemptId);
         return this.parseToPupilAttempt(response);
     }
 
     getPupilAttempts(): Array<PupilAttempt> {
-        var response = this.http.get<Array<PupilAttempt>>(this.url);
+        const response = this.http.get<Array<PupilAttempt>>(this.url);
         return this.parseToListPupilAttempts(response);
     }
 
-    savePupilAttempt(pupilAttempt: PupilAttempt) {
-        var response = this.http.post<PupilAttempt>(this.url + "/" + pupilAttempt.id + 
-                                "/" + pupilAttempt.attempt.id, pupilAttempt);
+    getPupilAttemptsById(pupilId: number): Array<PupilAttempt> {
+        const response = this.http.get<Array<PupilAttempt>>(this.url + '/' + pupilId);
+        return this.parseToListPupilAttempts(response);
+    }
+
+    recheckPupilAttempt(pupilId: number, attemptId: number): PupilAttempt {
+        const response = this.http.get<PupilAttempt>(this.url + '/recheck/' + pupilId + '/' + attemptId);
         return this.parseToPupilAttempt(response);
     }
 
-    updatePupilAttempt(pupilAttempt: PupilAttempt) {
-        var response;
-        this.http.put(this.url + "/" + pupilAttempt.id + 
-                                    "/" + pupilAttempt.attempt.id, pupilAttempt).subscribe(
-            item => {
-                response = item;
-            }, error => {
-                console.error(error);
-            });
-        return response;
+    savePupilAttempt(pupilAttempt: PupilAttempt): PupilAttempt {
+        const response = this.http.post<PupilAttempt>(this.url + '/' + pupilAttempt.id +
+                                '/' + pupilAttempt.attempt.id, pupilAttempt);
+        return this.parseToPupilAttempt(response);
+    }
+
+    updatePupilAttempt(pupilAttempt: PupilAttempt): PupilAttempt {
+        const response = this.http.put<PupilAttempt>(this.url + '/' + pupilAttempt.id +
+                                                  '/' + pupilAttempt.attempt.id, pupilAttempt);
+        return this.parseToPupilAttempt(response);
     }
 
     deletePupilAttempt(id: number) {
-        var response;
-        this.http.delete(this.url + "/" + id).subscribe(item => {
+        let response;
+        this.http.delete(this.url + '/' + id).subscribe(item => {
             response = item;
         }, error => {
             console.error(error);
@@ -53,7 +56,7 @@ import { PupilAttempt } from './pupilAttempt';
     }
 
     private parseToPupilAttempt(observable: Observable<PupilAttempt>): PupilAttempt {
-        var pupilAttempt = new PupilAttempt();
+        const pupilAttempt = new PupilAttempt();
         observable.subscribe(item => {
             pupilAttempt.id = item.id;
             pupilAttempt.attempt = item.attempt;
@@ -68,10 +71,10 @@ import { PupilAttempt } from './pupilAttempt';
     }
 
     private parseToListPupilAttempts(observable: Observable<Array<PupilAttempt>>): Array<PupilAttempt> {
-        var pupilAttempts = new Array<PupilAttempt>();
+        const pupilAttempts = new Array<PupilAttempt>();
         observable.subscribe(response => {
             response.map(item => {
-                var pupilAttempt = new PupilAttempt();
+                const pupilAttempt = new PupilAttempt();
                 pupilAttempt.id = item.id;
                 pupilAttempt.attempt = item.attempt;
                 pupilAttempt.mark = item.mark;
@@ -79,7 +82,7 @@ import { PupilAttempt } from './pupilAttempt';
                 pupilAttempt.pupil = item.pupil;
                 pupilAttempt.text = item.text;
                 pupilAttempts.push(pupilAttempt);
-            })
+            });
         }, error => {
             console.error(error);
         });
