@@ -1,13 +1,12 @@
 package com.vsu.by.app.education.attempt;
 
-import com.vsu.by.app.people.pupils.Pupil;
+import com.vsu.by.app.education.pupilattempt.PupilAttemptRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.sql.Date;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,6 +15,8 @@ public class AttemptService {
 
     @Autowired
     private AttemptRepository attemptRepository;
+    @Autowired
+    private PupilAttemptRepository pupilAttemptRepository;
 
     @Transactional(readOnly = true)
     public Optional<Attempt> getById(final Long id) {
