@@ -5,7 +5,6 @@ import com.vsu.by.app.education.task.dto.TaskDetailDto;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
-import java.sql.Date;
 import java.util.List;
 
 @Data
@@ -14,6 +13,6 @@ public class AttemptAddEditDto {
     private Long id;
     private TaskDetailDto task;
     private List<PupilAttemptInfoDto> pupilsAttempts;
-    private Date start;
-    private Date end; /**TODO FOR DIPLOM*/
+    private Long start;
+    private Long end;
 }
