@@ -1,11 +1,10 @@
 package com.vsu.by.app.education.attempt.dto;
 
-import com.vsu.by.app.education.pupilattempt.dto.PupilAttemptInfoDto;
+import com.vsu.by.app.education.pupilattempt.dto.PupilAttemptDto;
 import com.vsu.by.app.education.task.dto.TaskDetailDto;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
-import java.sql.Date;
 import java.util.List;
 
 @Data
@@ -13,7 +12,7 @@ import java.util.List;
 public class AttemptDetailDto {
     private Long id;
     private TaskDetailDto task;
-    private List<PupilAttemptInfoDto> pupilsAttempts;/**TODO????*/
-    private Date start;
-    private Date end; /**TODO FOR DIPLOM*/
+    private List<PupilAttemptDto> pupilsAttempts;
+    private Long start;
+    private Long end;
 }
