@@ -1,5 +1,8 @@
 package com.vsu.by.app.education.mistake;
 
+import com.vsu.by.app.education.pupilmistake.PupilMistake;
+import com.vsu.by.app.education.rule.Rule;
+import com.vsu.by.app.education.task.Task;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,9 +21,14 @@ public class MistakeService {
         return this.mistakeRepository.findById(id);
     }
 
+    @Transactional(readOnly = true)
+    public Mistake getMistakeByRule(final Rule rule) {
+        return this.mistakeRepository.findByRule(rule);
+    }
+
     @Transactional
-    public Mistake saveMistake(final Mistake error) {
-       return this.mistakeRepository.save(error);
+    public Mistake saveMistake(final Mistake mistake) {
+        return this.mistakeRepository.save(mistake);
     }
 
     @Transactional
@@ -29,8 +37,13 @@ public class MistakeService {
     }
 
     @Transactional
-    public Mistake updateMistake(final Mistake error) {
-        return this.mistakeRepository.save(error);
+    public Mistake updateMistake(final Mistake mistake) {
+        return this.mistakeRepository.save(mistake);
+    }
+
+    @Transactional(readOnly = true)
+    public List<Mistake> findAllByTask(final Task task) {
+        return this.mistakeRepository.findAllByTask(task);
     }
 
     @Transactional(readOnly = true)
