@@ -1,23 +1,26 @@
 package com.vsu.by.app.education.pupilattempt;
 
-import com.vsu.by.app.education.pupilattempt.dto.PupilAttemptAddEditDto;
 import com.vsu.by.app.education.pupilattempt.dto.PupilAttemptInfoDto;
 import com.vsu.by.app.education.pupilattempt.dto.PupilAttemptMapper;
+import com.vsu.by.app.education.pupilattempt.dto.PupilAttemptMinInfoDto;
+import com.vsu.by.app.education.pupilattempt.dto.PupilAttemptViewDto;
+import com.vsu.by.app.education.task.TaskService;
 import com.vsu.by.app.people.pupils.Pupil;
 import com.vsu.by.app.people.pupils.PupilService;
-import com.vsu.by.app.service.PupilAttemptChecker;
+import com.vsu.by.app.service.checking.PupilAttemptChecker;
+import com.vsu.by.app.service.text.TextProcessing;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("pupil-attempts")
+@RequestMapping("pupils-attempts")
 public class PupilAttemptController {
     @Autowired
     private PupilService pupilService;
@@ -27,17 +30,36 @@ public class PupilAttemptController {
     private PupilAttemptMapper pupilAttemptMapper;
     @Autowired
     private PupilAttemptChecker pupilAttemptChecker;
+    @Autowired
+    private TaskService taskService;
+    @Autowired
+    private TextProcessing textProcessing;
+
+    @GetMapping
+    public ResponseEntity<List<PupilAttemptMinInfoDto>> getPupilsAttempts() {
+        List<PupilAttempt> pupilAttempts = this.pupilAttemptService.findAll();
+        return new ResponseEntity<>(
+                this.pupilAttemptMapper.toPupilAttemptMinInfoDto(pupilAttempts),
+                HttpStatus.OK);
+    }
+
+    @GetMapping("/{pupilId}")
+    public ResponseEntity<List<PupilAttemptInfoDto>> getPupilAttemptsById(@PathVariable("pupilId") Long pupilId) {
+        List<PupilAttempt> pupilAttempts = this.pupilAttemptService.findAllByPupilId(pupilId);
+        return new ResponseEntity<>(
+                this.pupilAttemptMapper.toPupilAttemptInfoDto(pupilAttempts),
+                HttpStatus.OK);
+    }
 
     @GetMapping("/{pupilId}/{attemptId}")
-    public ResponseEntity<PupilAttemptInfoDto> getPupilAttempt(@PathVariable("pupilId") Long pupilId,
-                                                               @PathVariable("attemptId") Long attemptId,
-                                                               Model model) {
+    public ResponseEntity<PupilAttemptViewDto> getPupilAttempt(@PathVariable("pupilId") Long pupilId,
+                                                               @PathVariable("attemptId") Long attemptId) {
         Optional<Pupil> pupil = this.pupilService.getPupil(pupilId);
         if (pupil.isPresent()) {
             PupilAttempt pupilAttempt = this.pupilAttemptService.getByAttemptAndByPupil(attemptId, pupil.get());
             if (Objects.nonNull(pupilAttempt)) {
                 return new ResponseEntity<>(
-                        this.pupilAttemptMapper.toPupilAttemptInfoDto(pupilAttempt),
+                        this.pupilAttemptMapper.toPupilAttemptViewDto(pupilAttempt),
                         HttpStatus.OK);
             } else {
                 /**TODO EXCEPTION*/
@@ -51,23 +73,66 @@ public class PupilAttemptController {
 
     /**TODO PutMapping?*/
     @PutMapping("/{pupilId}/{attemptId}")
-    public ResponseEntity<String> updatePupilAttempt(@PathVariable("pupilId") Long pupilId,
+    public ResponseEntity<PupilAttemptInfoDto> updatePupilAttempt(@PathVariable("pupilId") Long pupilId,
                                      @PathVariable("attemptId") Long attemptId,
-                                     @RequestBody PupilAttemptAddEditDto pupilAttemptAddEditDto) {
-        this.pupilAttemptService.updatePupilAttempt(
-                this.pupilAttemptMapper.fromPupilAttemptAddEditDto(pupilAttemptAddEditDto));
-        return new ResponseEntity<>("Pupil attempt was updated", HttpStatus.OK);
+                                     @RequestBody PupilAttemptInfoDto pupilAttemptDto) {
+        PupilAttempt pupilAttempt =
+                this.pupilAttemptMapper.fromPupilAttemptInfoDto(pupilAttemptDto);
+        pupilAttempt.setText(this.textProcessing.processText(pupilAttempt.getText()));
+        this.initializeNullFields(pupilAttempt);
+        pupilAttempt = this.pupilAttemptService.updatePupilAttempt(pupilAttempt);
+        return new ResponseEntity<>(
+                this.pupilAttemptMapper.toPupilAttemptInfoDto(pupilAttempt),
+                HttpStatus.ACCEPTED);
     }
 
     /**TODO POSSIBLE OR EVERYTHING OK*/
     @PostMapping("/{pupilId}/{attemptId}")
-    public ResponseEntity<String> savePupilAttempt(@PathVariable("pupilId") Long pupilId,
+    public ResponseEntity<PupilAttemptInfoDto> savePupilAttempt(@PathVariable("pupilId") Long pupilId,
                                    @PathVariable("attemptId") Long attemptId,
-                                   @RequestBody PupilAttemptAddEditDto pupilAttemptAddEditDto) {
+                                   @RequestBody PupilAttemptInfoDto pupilAttemptDto) {
         PupilAttempt pupilAttempt =
-                this.pupilAttemptMapper.fromPupilAttemptAddEditDto(pupilAttemptAddEditDto);
+                this.pupilAttemptMapper.fromPupilAttemptInfoDto(pupilAttemptDto);
+        pupilAttempt.setText(this.textProcessing.processText(pupilAttempt.getText()));
+        this.initializeNullFields(pupilAttempt);
+        this.checkPupilAttempt(pupilAttempt);
+        pupilAttempt = this.pupilAttemptService.savePupilAttempt(pupilAttempt);
+        return new ResponseEntity<>(
+                this.pupilAttemptMapper.toPupilAttemptInfoDto(pupilAttempt),
+                HttpStatus.ACCEPTED);
+    }
+
+    @GetMapping("/recheck/{pupilId}/{attemptId}")
+    public ResponseEntity<PupilAttemptViewDto> recheckPupilAttempt(@PathVariable("pupilId") Long pupilId,
+                                                                   @PathVariable("attemptId") Long attemptId) {
+        Optional<Pupil> pupil = this.pupilService.getPupil(pupilId);
+        if (pupil.isPresent()) {
+            PupilAttempt pupilAttempt =
+                    this.pupilAttemptService.getByAttemptAndByPupil(attemptId, pupil.get());
+            if (Objects.nonNull(pupilAttempt)) {
+                this.checkPupilAttempt(pupilAttempt);
+                pupilAttempt = this.pupilAttemptService.savePupilAttempt(pupilAttempt);
+                return new ResponseEntity<>(
+                        this.pupilAttemptMapper.toPupilAttemptViewDto(pupilAttempt),
+                        HttpStatus.OK);
+            } else {
+                /**TODO EXCEPTION*/
+                throw new NoSuchElementException("Такой попытки не существует");
+            }
+        } else {
+            /**TODO EXCEPTION*/
+            throw new NoSuchElementException("Такого ученика не существует");
+        }
+    }
+
+    /**TODO А нахуя?*/
+    private void initializeNullFields(PupilAttempt pupilAttempt) {
+        pupilAttempt.setPupil(this.pupilService.getPupil(pupilAttempt.getPupil().getId()).get());
+        pupilAttempt.getAttempt().setTask(
+                this.taskService.getTask(pupilAttempt.getAttempt().getTask().getId()).get());
+    }
+
+    private void checkPupilAttempt(PupilAttempt pupilAttempt) {
         this.pupilAttemptChecker.checkPupilAttempt(pupilAttempt);
-        this.pupilAttemptService.savePupilAttempt(pupilAttempt);
-        return new ResponseEntity<>("Pupil attempt was saved", HttpStatus.OK);
     }
 }
