@@ -7,10 +7,8 @@ import javax.persistence.*;
 @Data
 @Entity
 public class Rule {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "id_seq")
-    @SequenceGenerator(name="id_seq", sequenceName = "rule_id", allocationSize = 1)
-    private Long id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id private Long id;
     @Column(length = 164, unique = true)
     private String name;
     private String description;
