@@ -7,10 +7,9 @@ import javax.persistence.*;
 @Data
 @Entity
 public class Subject {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "id_seq")
-    @SequenceGenerator(name="id_seq", sequenceName = "subject_id", allocationSize = 1)
-    private Long id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    //@SequenceGenerator(name="id_seq", sequenceName = "subject_id", allocationSize = 1)
+    @Id private Long id;
     @Column(nullable = false, length = 32)
     private String name;
 }
