@@ -8,15 +8,13 @@ import javax.persistence.*;
 @Data
 @Entity
 public class Task {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "id_seq")
-    @SequenceGenerator(name="id_seq", sequenceName = "task_id", allocationSize = 1)
-    private Long id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id private Long id;
     @Column(length = 64, nullable = false)
     private String name;
     @Embedded
     private TaskInfo info;
-    @OneToOne(cascade = {CascadeType.ALL})
+    @OneToOne(cascade = {CascadeType.MERGE})
     private TaskType type;
     /**TODO FOR DIPLOM*/
     private String taskText;
