@@ -2,11 +2,13 @@ package com.vsu.by.app.education.task;
 
 import com.vsu.by.app.education.subject.Subject;
 import com.vsu.by.app.people.teacher.Teacher;
-import com.vsu.by.app.people.user.User;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToOne;
 
 @Data
 @RequiredArgsConstructor
@@ -15,7 +17,8 @@ public class TaskInfo {
     @Column(length = 128, nullable = false)
     private String description;
     @OneToOne(cascade = {CascadeType.MERGE, CascadeType.REFRESH})
-    private Teacher uploader;
+    private Teacher teacher;
     @OneToOne(cascade = {CascadeType.REFRESH})
     private Subject subject;
+
 }
