@@ -10,8 +10,9 @@ import javax.persistence.*;
 @Data
 @RequiredArgsConstructor
 @Entity
+@PrimaryKeyJoinColumn(name = "pupil_id")
 public class Pupil extends User{
-    @ManyToOne(optional = false)
-    @JoinColumn(insertable = false, updatable = false, name = "id")
+    @ManyToOne(optional = false, cascade = {CascadeType.MERGE})
+    @JoinColumn(name = "group_id")
     private Group group;
 }
