@@ -1,5 +1,7 @@
 package com.vsu.by.app.people.pupils.dto;
 
+import com.vsu.by.app.people.groups.dto.GroupInfoDto;
+import com.vsu.by.app.role.dto.RoleDto;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
@@ -7,9 +9,10 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class PupilInfoDto {
     private Long id;
+    private String login;
     private String name;
     private String surname;
     private String patronymic;
-    private String faculty;
-    private int groupNo;
+    private GroupInfoDto group;
+    private RoleDto role;
 }
